@Scheduled测试

+++ src/main/java/com/spring/boot/test/project/dao/SpringJdbcDao.java
@@ -15,9 +15,9 @@ public interface SpringJdbcDao {
 
     int update(SpringJdbcEntity springJdbcEntity);
 
-    int delete(int id);
+    int delete(Long id);
 
-    SpringJdbcEntity findAccountById(int id);
+    SpringJdbcEntity findAccountById(Long id);
 
     List<SpringJdbcEntity> findAccountList();
 }
